# swagger.yml
host: 127.0.0.1:3000
info:
  description: Todo Foo  bar
  title: Todo list app
  version: "0.1"
consumes:
  - application/json
#  - application/xml
produces:
  - application/json
#  - application/xml
schemes:
  - http
paths:
  /:
    post:
      operationId: createTodoItem
      tags:
        - todo
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/item"
      responses:
        201:
          description: Created
          schema:
            $ref: "#/definitions/item"
        default:
          description: error occured
          schema:
            $ref: "#/definitions/error"


    get:
      operationId: fetchTodoItems
      tags:
        - todo

      responses:
        200:
          description: fetch todo items
          schema:
            type: array
            items:
              $ref: "#/definitions/item"
        default:
          description: error occured
          schema:
            $ref: "#/definitions/error"






  /{id}:
    parameters:
      - type: integer
        format: int64
        name: id
        in: path
        required: true

    put:
      operationId: replaceTodoItem
      tags:
        - todo
      parameters:
        - name: body
          in: body
          schema:
            $ref: "#/definitions/item"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/item"
        404:
          description: Not Found
          schema:
            $ref: "#/definitions/error"
        default:
          description: error
          schema:
            $ref: "#/definitions/error"

    delete:
      operationId: deleteTodoItem
      tags:
        - todo
      responses:
        204:
          description: Deleted
        404:
          description: Not Found
          schema:
            $ref: "#/definitions/error"
        default:
          description: error
          schema:
            $ref: "#/definitions/error"
    get:
      operationId: fetchTodoItem
      tags:
        - todo
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/item"
        404:
          description: Not Found
          schema:
            $ref: "#/definitions/error"
        default:
          description: error
          schema:
            $ref: "#/definitions/error"

definitions:
  item:
    type: object
    required:
      - description
      - id
    properties:
      id:
        type: integer
        format: int64
        readOnly: true
      description:
        type: string
        minLength: 1
      completed:
        type: boolean



  error:
    type: object
    required:
      - message
    properties:
      message:
        type: string
swagger: "2.0"